\name{NormalLaplace-package}
\alias{NormalLaplace-package}
\alias{NormalLaplace}
\docType{package}
\title{
The Package `NormalLaplace': Summary Information
}
\description{
  
This package provides a collection of functions for Normal Laplace
distributions. Functions are provided for the density function,
distribution function, quantiles and random number generation.The mean,
variance, skewness and kurtosis of a given Normal Laplace distribution
are given by \code{nlMean}, \code{nlVar}, \code{nlSkew}, and
\code{nlKurt} respectively.


}

\author{
David Scott \email{d.scott@auckland.ac.nz}, Jason Shicong Fu

Maintainer: David Scott <d.scott@auckland.ac.nz>

}
\references{
  William J. Reed. (2006) The Normal-Laplace Distribution and Its
  Relatives. In \emph{Advances in Distribution Theory, Order Statistics
  and Inference}, pp. 61--74.  \enc{Birkhäuser}{Birkhauser},
  Boston.
}

\keyword{package}
\seealso{
  \code{\link{dnl}}, \code{\link{millsR}},
  \code{\link{NormalLaplaceMeanVar}}
}



